+++ CHANGELOG.md
@@ -11,6 +11,16 @@ for further information about branching and tagging conventions.
 #### Added
 - Form validation of doc types in limits form
 
+#### Fixed
+- A bug with possibility select start date after end date
+  in SaleList and PollList
+
+#### Changed
+- InputDateField
+
+#### Removed
+- vue-flatpickr-component node module
+
 ## [1.13.0-rc.0] - 2020-03-25
 #### Fixed
 - Details in createChangeRoleRequest at unblock user